Move util functions to calculate_block_cone and projection
